Add mean function; edits to variance function
I created a new "mean" function in src/survey/_descriptive_stats.py. I
also created the relevant tests of the mean function in
tests/test_descriptive_stats.py. As noted below, I checked that I could
calculate the standard error for this function. Hopefully this sets up
teh template so other functions can be easily included in this package.

I made some edits to the src/survey/_agg.py function, specifically
creating a second aggregator (besides _count) that can be used to apply
different functions to grouped variables (so you can, say, calculate the
mean of var 1 over each of the values of var 2).

I made a lot of edits to the _variance function so it can be used to
find the standard error of a function (like the mean), both on its own
and when grouped by another value.