Edit weighted mean function to fillna()/var tests
Before, the dot product in the weighted mean function did not have
a .fillna() on the weights; this resulted in the answer being nan
if any weights were missing.

I also did minor edits to the variance tests. On the last commit,
I changed the code so the index properly renames when missing; now
values that are missing are named "NA". I updated the indices in the
variance tests to reflect that.